Rename mt19937_rng to mt19937
