+++ mk/tests.mk
@@ -11,7 +11,7 @@ CACHE_TEST_OBJS := \
 
 MAP_TEST_OBJS := \
 	test-map.o \
-	mt19937_rng.o
+	mt19937.o
 
 CACHE_TEST_OBJS := $(addprefix $(CACHE_TEST_OUTDIR)/, $(CACHE_TEST_OBJS)) \
 		   $(OUT)/cache.o $(OUT)/mpool.o
